+++ native/src/BackendRenderTarget.cc
@@ -15,6 +15,17 @@ extern "C" JNIEXPORT jlong JNICALL Java_org_jetbrains_skija_BackendRenderTarget_
 extern "C" JNIEXPORT jlong JNICALL Java_org_jetbrains_skija_BackendRenderTarget__1nMakeGL
   (JNIEnv* env, jclass jclass, jint width, jint height, jint sampleCnt, jint stencilBits, jint fbId, jint fbFormat) {
     GrGLFramebufferInfo glInfo = { static_cast<unsigned int>(fbId), static_cast<unsigned int>(fbFormat) };
-    GrBackendRenderTarget* obj = new GrBackendRenderTarget(width, height, sampleCnt, stencilBits, glInfo);
-    return reinterpret_cast<jlong>(obj);
+    GrBackendRenderTarget* instance = new GrBackendRenderTarget(width, height, sampleCnt, stencilBits, glInfo);
+    return reinterpret_cast<jlong>(instance);
 }
+
+#ifdef SK_METAL
+extern "C" JNIEXPORT jlong JNICALL Java_org_jetbrains_skija_BackendRenderTarget__1nMakeMetal
+  (JNIEnv* env, jclass jclass, jint width, jint height, jlong texturePtr) {
+    GrMTLHandle texture = reinterpret_cast<GrMTLHandle>(static_cast<uintptr_t>(texturePtr));
+    GrMtlTextureInfo fbInfo;
+    fbInfo.fTexture.retain(texture);
+    GrBackendRenderTarget* instance = new GrBackendRenderTarget(width, height, fbInfo);
+    return reinterpret_cast<jlong>(instance);
+}
+#endif