+++ shared/src/main/java/org/jetbrains/skija/BackendRenderTarget.java
@@ -6,9 +6,16 @@ import org.jetbrains.skija.impl.*;
 public class BackendRenderTarget extends Managed {
     static { Library.staticLoad(); }
     
+    @NotNull @Contract("_, _, _, _, _, _ -> new")
     public static BackendRenderTarget makeGL(int width, int height, int sampleCnt, int stencilBits, int fbId, int fbFormat) {
         Stats.onNativeCall();
         return new BackendRenderTarget(_nMakeGL(width, height, sampleCnt, stencilBits, fbId, fbFormat));
+    }
+
+    @NotNull @Contract("_, _, _ -> new")
+    public static BackendRenderTarget makeMetal(int width, int height, long texturePtr) {
+        Stats.onNativeCall();
+        return new BackendRenderTarget(_nMakeMetal(width, height, texturePtr));
     }
 
     @ApiStatus.Internal
@@ -21,6 +28,7 @@ public class BackendRenderTarget extends Managed {
         public static final long PTR = _nGetFinalizer();
     }
 
-    public static native long _nGetFinalizer();
-    public static native long _nMakeGL(int width, int height, int sampleCnt, int stencilBits, int fbId, int fbFormat);
+    @ApiStatus.Internal public static native long _nGetFinalizer();
+    @ApiStatus.Internal public static native long _nMakeGL(int width, int height, int sampleCnt, int stencilBits, int fbId, int fbFormat);
+    @ApiStatus.Internal public static native long _nMakeMetal(int width, int height, long texturePtr);
 }